Use stunnel for encryption.
Basic idea is to start an stunnel daemon for both the client and the
daemon.  The client/daemon talk to the stunnel instances which talk
directly to each other.  Handled by writing an stunnel configuration
file on the fly and running it as a child process.

Test cases have been updated.

TODO (off the top of my head):
    - Effectively breaks unencrypted communication when enabled.
    - #ifdef mess.